- Use new protobufs, so had to change timestamp handling to be uint32. Matching of JSON in tests also updated to check for special case where expected is string (with SECAGO or whatever in it) and we receive float64 - Index compression (run length encoding incrementing values in list) now works both as encode and decode, and moved into core indexcompression package, used by scan data request index lists still but now also by ROI pixel and scan entry index lists - Implemented full set of ROI CRUD operations, wrote tests for it based on element sets. Found some bugs, fixed in element sets too related to modified time not being set. Also standardised validation of element sets and ROIs using helper funcs written since element sets written - V3 migration tool can now be given a list of datasets to include, to make quick test data imports quicker, applies compression to location+pixel indexes for ROIs on the fly

+++ api/ws/handlers/user-group-joining.go
@@ -40,7 +40,7 @@ func HandleUserGroupJoinReq(req *protos.UserGroupJoinReq, hctx wsHelpers.Handler
 		Id:             reqId,
 		UserId:         hctx.SessUser.User.Id,
 		JoinGroupId:    req.GroupId,
-		CreatedUnixSec: uint64(hctx.Svcs.TimeStamper.GetTimeNowSec()),
+		CreatedUnixSec: uint32(hctx.Svcs.TimeStamper.GetTimeNowSec()),
 		AsMember:       req.AsMember,
 	}
